+++ test/Interop/Cxx/class/Inputs/simple-structs.h
@@ -52,38 +52,4 @@ public:
   Outer() : privStruct(1, 2, 3, 4, 5, 6), publStruct(7, 8, 9, 10, 11, 12) {}
 };
 
-struct ImmortalFRT {
-private:
-  int priv = 1;
-
-public:
-  int publ = 2;
-} __attribute__((swift_attr("import_reference")))
-__attribute__((swift_attr("retain:immortal")))
-__attribute__((swift_attr("release:immortal")));
-
-struct FRTCustomStringConvertible {
-public:
-private:
-  int priv = 1;
-
-public:
-  int publ = 2;
-} __attribute__((swift_attr("import_reference")))
-__attribute__((swift_attr("retain:immortal")))
-__attribute__((swift_attr("release:immortal")));
-
-struct FRType {
-private:
-  int priv = 1;
-
-public:
-  int publ = 2;
-} __attribute__((swift_attr("import_reference")))
-__attribute__((swift_attr("retain:retain")))
-__attribute__((swift_attr("release:release")));
-
-void retain(FRType *v) {};
-void release(FRType *v) {};
-
 #endif
